requirements.txt
Django==4.2.3
djangorestframework==3.14.0
drf-spectacular==0.26.3
PyYAML==6.0.1